refactor: extract organization members panel
